Fix spdlog build
When logging was enabled, there were numerous compiler errors of the
sort: "destructor called on non-final that has virtual functions but
non-virtual destructor", which appear to be fixed by marking all final
classes as final.

libfmt also supposedly requires you to register your types so it knows
how to format them (https://fmt.dev/latest/api/#udt), but this was not
done.  The solution is to add the following for any class to be
formatted:

```cpp
class MyFormattableClass {};

#ifdef UAGENT_LOGGER_PROFILE
#include <spdlog/fmt/bundled/ostream.h>
template <> struct fmt::formatter<MyFormattableClass> : ostream_formatter {};
#endif
```